backend/requirements.txt
aiohttp
python-dotenv
websockets
openai
azure-identity
